Backend/Pong/requirements.txt
django
pillow
requests
django-allauth
psycopg2
djangorestframework
django-filter
djangorestframework-simplejwt
django-cors-headers
daphne
channels